with_simulate for multiple runs
with_simulate replaces with_vars_run.

with_simulate allows also runs without a parameter sweep. And always
writes to disk, rather than taking a function to specify an action.

This decouples the actions for the running, through the disk, which
seems like a reasonable for the use cases we have in mind.

with_simulate returns a tibble with the input variables and the
location of the results file.